// src/types/site.ts
export interface NavLink {
  label: string
  href: string
}

export interface SystemCard {
  n: string
  verb: string
  noun: string
  product: string
  img: string
  h: string
}

export interface StatItem {
  v: number
  suf: string
  l: string
}

export interface ProcessStep {
  n: string
  title: string
  description: string
}

export interface ContactItem {
  label: string
  value: string
  href: string
}